Huge Revamp
Basically refactored the whole thing to take in function arguments rather than global variables. This makes it possible to have default values when calling main(), which is a huge improvement on the previous experience for non-technical people - we can have a number of preset games to run, a much more readable editable section, and a simple guide to running the program. This sets us up really well for future alterations. Thanks, OOD :)